Upgrade PySide6 to 6.6.2 to fix macOS crashing bug
This fixes the crash on macOS when clicking the Refresh button, if a
background tool that uses the Accessibility APIs is running, such as
Magnet (a window manager).

To minimize regression risks, I bumped the release of PySide6 just
enough to contain the fix for the bug. As it turns out, it was fixed in
the patch version just above the one we were using.

See https://bugreports.qt.io/browse/QTBUG-120469