Switch back to maintained bitarray package
The `bitarray-hardbyte` package on PyPI stopped being updated back in
2022 when its packaging improvements were reintegrated into the main
`bitarray` project, see
https://github.com/ilanschnell/bitarray/issues/165#issuecomment-1054075870

There doesn't seem to be a reason to depend on the `bitarray-hardbyte` package
which is no longer maintained. We now depend on the current release of
`bitarray` (2.9.2) which among other things initializes new arrays to
zero, which could help avoid some obscure bugs. This guaranteed
initialization is now used in two places where a temporary `[False,
False...]` array was previously constructed just to initialize a
`bitarray`.