engine_fsm: avoid hanging on test exit
The engine would get stuck in RUN_TEST state waiting for a message
from the firmware that might never come. And there would be no way
to exit that waiting loop from the UI.

By returning to the main loop, we give the UI a chance to exit test
mode.