Copy issue template and contribution guide from main branch
Since the default branch for the repo has been switched from `main` to `1.0.0-dev` the issue template is no longer applied and the contribution guide no longer appears.

This copies the two files from the `main` branch — I also fixed the name of `main` branch (was `master`) in `CONTRIBUTING.md`.