Remove requirement for separate Qt installation
The only reason to require a standalone Qt installation was the
`lrelease` tool used to convert the translation files.

It turns out PySide6, which we install through `pip`, already includes
this tool under the name `pyside6-lrelease`.

This changes the call from `lrelease` to `pyside6-lrelease`, and removes
the installation of Qt from the GitHub Actions workflow and from the
development setup instructions, except on Linux where we install Qt tools
just to bring in the required system libraries.