Make scene.refresh work without an image (#691)
* Make scene.refresh work without an image

There was an assumption that scene.refresh() would not be called unless
an image had been loaded. But there are several ways to have it called,
among which:
 - if the "Default Knit Side Image" setting is checked, the initial
   `reverse()` call will call `refresh()`;
 - whenever the preferences panel is closed, `refresh()` is called, even
   if an image has not yet been loaded.

It seems simpler to make `refresh()` able to work without an image.

A visible change is that the "machine" (the orange and green bars) will
be visible from the start, which I think is an improvement.

* scene: fix accessing property to soon in __init__

The call to `reverse` that happens in `Scene.__init__` triggers a `refresh` that in turn accesses the `zoom` property. This creates an exception if the property hasn't been initialized yet.