Fix hardware test dialog (#693)
* communication: sync Token values with cmd list from firmware

* hw_test: remove extra argument in signal handler

* hw_test: fix command list

* hw_test mock: make help message match firmware

* hw_test: fix commands

* hw_test mock: fix commands

* engine_fsm: avoid hanging on test exit

The engine would get stuck in RUN_TEST state waiting for a message
from the firmware that might never come. And there would be no way
to exit that waiting loop from the UI.

By returning to the main loop, we give the UI a chance to exit test
mode.

* engine_fsm: fix entering test mode

The previous code waited for an `indState` message but the current
firmware doesn't seem to be sending this. So the REQUEST_TEST state
only has to emit `reqTest` and transition to CONFIRM_TEST. Maybe the
state could be removed entirely.

* communication: use SLIP encoding in write_API6

The firmware definitely expects all commands to be wrapped in SLIP
packets.
The write_API6 method is only used by the hardware test commands,
this is why this oversight went undetected.

* engine_fsm: don't loop from FINISHED to CONNECT

There was a race condition caused by `HardwareTestDialog.reject` trying to end testing
by forcing `__control.state` to `FINISHED` from the UI thread.
If this occurred as the thread was about to enter `Control.operate`, the
engine FSM would immediately jump from state `FINISHED` to state
`CONNECT`, so `Engine.run` would never
get a chance to notice the `FINISHED` state and exit its processing loop.

This was causing random hangs when exiting the hardware test dialog.

By making `FINISHED` a "dead-end" state, we make sure `Engine.run` eventually exits its
loop, and the thread terminates.

* engine_fsm: remove call to control.timer

There is no such property. The statement was always raising an error that was silently ignored.

* hw_test: remove commented out code