DAK file reading: don't assume UTF-8 encoding
When extracting strings from a DAK (.stp or .pat) format file, we shouldn't assume that
these strings are encoded using UTF-8.
In particular, it appears localized color names are not encoded using UTF-8, and
trying to decode them as UTF-8 causes pattern loading to fail.
Another place where strings are read from the files is as part of the .stp decryption
process. In that case, it seems safer to not try to interpret those byte sequences at all.

This commit changes the strings to Python bytestrings, so we don't need to guess at the
encoding that DAK actually uses for color names, and we can be sure the bytes in the
encryption keys are unmodified.