version_checker: don't delete network reply in finished handler
The Qt docs warn against deleting a QNetworkReply while the
`finished` handler is running (https://doc.qt.io/qt-6/qnetworkreply.html).

We're not explicitly deleting it here, just setting the Python reference
to `None`, but that might be enough to cause the object to be
destroyed, explaining the crash we're seeing on Windows.

Instead we set the option on QNetworkAccessManager to auto-delete
replies.